set map handle for tiled add layer add physique add hide system remove xml custom map

+++ core/src/com/kikijoli/ville/drawable/entite/npc/Guard.java
@@ -28,7 +28,7 @@ public class Guard extends Entite implements IBusiness {
 
     public PointLight vision;
 
-    public Guard(int srcX, int srcY) {
+    public Guard(float srcX, float srcY) {
         super(GUARD, srcX, srcY);
         this.buisiness = this.getDefault();
         this.vision = new PointLight(Tmap.getRay(), 40, Color.BLACK, 500, this.getX(), this.getY());
